[ci] Remove clone_depth from appveyor.yml
clone_depth: 1 means we only need to clone the last 1 revision. Sounds
good! Except, it's not the last 1 starting from whatever revision we
want to test, it's the last 1 starting from the current HEAD of the
repo. So clone_depth: 1 means that we might not get the revision we
wanted at all.

Our repo is tiny enough that this optimization probably isn't even
worthwhile, so I just removed it.