Yet another attempt to make test_ki_wakes_us_up reliable on Windows
It failed yet again, even with all our countermeasures:
  https://ci.appveyor.com/project/njsmith/trio/build/1.0.341/job/r356mgh7mi103lip

Current theory: the two ki_self()'s are running, and each is
synchronously running the C-level signal handler in the kill_soon
thread (because ki_self() uses raise() and I think it works like
that). And then the main thread is taking the lock, to try to
guarantee that it runs some inside the test after *both* ki_self()'s,
so that any Python-level handlers that are going to run, run inside
the test.

But, maybe somehow the main thread is not noticing that the C-level
handler has requested that the Python-level handler be run, because
PyEval_EvalFrameEx only checks every once in a while? I'm skeptical
because it's supposed to notice promptly after the C-level signal
handler runs, but maybe somehow this isn't working? So this adds an
explicit call to PyErr_CheckSignals. Except of course we can't write
it like that, because that's a C function. Instead we call repr(),
which should trigger it.

We're a long way down the rabbit hole.