test_multierror: Use IPython's code_to_run instead of exec_lines
The reason the tests "freeze" on AppVeyor should have been that
`exec_lines` runs the lines at IPython startup and then leaves the
python shell open, meaning that the process would never exit.  This
should no longer be the case with `code_to_run`, so remove the skip
marker as well.

Signed-off-by: Johannes Löthberg <johannes@kyriasis.com>