Simplify and speed up ParkingLot
Use OrderedDict instead of SortedDict, and don't make park() pay for
repark().

See gh-272 for context. This doesn't make ParkingLot as fast as a
version that just holds a set of tasks and does

  def unpark_all(self):
      while self._waiters:
          reschedule(self._waiters.pop())

  def unpark(self):
      if self._waiters:
          reschedule(self._waiters.pop())

but it's about half-way in between the old version and this minimal
version on the broker microbenchmark.