Avoid infinite recursion in ModuleWithDeprecations.__getattr__
This isn't really needed currently (we'd know, the symptom is a
RecursionError at import time), but it's a genuine bug fix for a
problem I ran into while implementing gh-316. And since that PR seems
to be stalled at the moment, I want to get this in so it doesn't get
forgotten.