Remove is_trio_socket
As pointed out in:

  https://github.com/python-trio/trio/issues/170#issuecomment-327025129

there are advantages to having the "is this a trio socket" check be
spelled using 'isinstance'. This commit un-deprecates the
trio.socket.SocketType name and makes it an abstract class that
concrete socket implementations should inherit from, and then gets rid
of is_trio_socket since it's now unnecessary.