Attempt to fix MacOS failures on travis by increasing sleep lengths
Our tests that sleep actually sleeps for the specified time started
failing on Travis's MacOS builds -- we were requesting to sleep for
0.05 seconds, and woke up after ~0.11 seconds, which is a large enough
overrun to break our tests. The problems started happening without any
changes to our code, so pretty sure this is a Travis issue, not a Trio
issue.

Still, I guess we need to make these tests more robust. The approach
used here is to bump up the sleep time massively from 0.05 seconds to
1.00 seconds, and then mark the relevant tests as @slow. This adds ~5
seconds to the --run-slow test suite, which is something like a 50%
increase on my laptop (but proportionally less if gathering coverage
etc). We can try bumping the number down again later if we need to
speed things up.