Add a Stream adapter for raw file handles to trio.hazmat
ReadFDStream and WriteFDStream are useful for attaching to raw file descriptors,
e.g. those passed to the parent side of a subprocess's stdin/out/err or
.
They are probably not suitable for accessing "real" stdin/out.

TODO: There aren't yet any filter classes for reading line-by-line,
translating newlines, or character set translation.
As of this patch, these handlers destroy a Python io.* handle
if you use it with them.
That's (part of) the reason why they're in hazmat.