Add "trio.hazmat.wait_priority(fd)".
This function is only available for epoll. It uses the EPOLLPRI mode to
trigger on an "exceptional condition", whatever that is with respect to the
given file descriptor.

Among other uses, this flag can be used to wait for changes on GPIO inputs,
or for mode changes of a PTY slave device (when used on the master).