Tweak backlog test to be slightly more strict
The point of this test is to ask trio to create sockets with two
different backlogs, and then check that they actually have different
backlogs, to make sure that trio is paying attention to our request.
But the two ranges we checked for overlapped, so that it was
theoretically possible that if the backlog was always exactly 10, then
the test could pass, even if trio was ignoring the backlog argument
entirely.

This is not very likely but I noticed it while I was in here fixing
gh-522 and it bugs me.