Use ABCs instead of inspect to detect generators/coroutines
This is important because Cython has its own generator/coroutine
types that are interchangeable with the ones built into the
interpreter, but not (always) detectable by the inspect module.

There's some more checks in trio/_core/_ki.py that still use the
inspect module. Unfortunately, they're using the inspect.is*function
forms, which are not easily replaceable by ABC checks. Fortunately,
usage of @enable_ki_protection and @disable_ki_protection are very
rare, so hopefully this won't affect too much?

Unfortunately, this is not sufficient to support Cython – we also need
to do something about how we handle KI protection. See gh-550 for
details. But this is still a good idea on its own.