Give more details when the init task crashes
This should never happen in normal use, but sometimes there are bugs
that can cause the init task to crash. (For example, I hit this while
trying to debug gh-550.) Previously, this would trigger an assertion
failure, and the original exception would be lost. Let's preserve the
original exception to make debugging easier.