Add CapacityLimiter.UNLIMITED
Sometimes we have an API that wants a CapacityLimiter (like
run_sync_soon_in_worker_thread), and we want to have unlimited
capacity (for example, when putting a long-running blocking call into
a background thread, like WaitForMultipleObjects or waitpid). This can
be done by writing a little stub class that implements the
CapacityLimiter interface, but doing this all the time is annoying.
Here's a full-fledged version, available whenever you need it.