s/result/outcome/ in _run.py
The original name for Outcome was Result, but it got changed when we
moved it to an external library. (The PyPI name was already taken.)
All the code in _run.py still referred to outcomes as results. This
worked, but was pretty confusing to try to read.

This commit is just a global search/replace of "result" with "outcome"
in _run.py (with a bit of care to avoid breaking things).