Generalize namespace static analysis tests
This switches to a single parametrized test that covers all of Trio's
public namespaces, and if it fails, then it prints a nice readable
list of the missing symbols.

Also adds the symbols that were missing from trio.socket's static
import list on my laptop, and adds some comments to trio.socket while
I'm at it.