Add trio's own dependencies to *-requirements.in
Repeating the dependencies like this is a little gross, but it enables
them to be pinned, and makes it so you can actually test trio after
installing test-requirements.txt. Plus, if they do get out-of-sync
with setup.py, then nothing terrible will happen, we'll just be in a
similar position to what we were before this patch. So I guess it's an
improvement.

Also ran pip-compile to regenerate *-requiresments.txt.