Add trio.hazmat.batch_cancellations()
This generalizes the "cancel multiple scopes without waking up the tasks, then wake up all the tasks at once"
pattern that was previously a special case for handling cancellations on a deadline. It will simplify the
implementation of #886, and exposing it as part of hazmat lets users write their own cancellation
abstractions that work the way the native ones do.