Attempt to recover from cancel scope nesting violations
When exiting an outer cancel scope before some inner cancel scope nested inside it, Trio would previously raise a confusing chain of internal errors.
One common source of this misbehavior was an async generator that yielded within a cancel scope.
Now the mis-nesting condition will be noticed and in many cases the user will receive a single RuntimeError with a useful traceback.
This is a best-effort feature, and it's still possible to fool Trio if you try hard enough, such as by closing a nursery's cancel scope
from within some child task of the nursery.

Fixes #882. Relevant to #1056, #264.