Make testing instructions less generic
In `contributing.rst`, the instructions for running tests locally appear to have been cut-and-pasted from the pytest manual, or something like that, and it's not clear exactly what to type.  Replacing "project" with "trio" in two places fixes this.  In particular, I have verified that `pytest trio` is the correct command to run the test suite once everything is set up.

Also use a code-block instead of a bulleted list of verbatim strings.