Stop closing stdin in tests
When FdStream(0) becomes unreachable, _FdHolder is garbage-collected and
its __del__ is called, which closes the file descriptor. For some
reason, with the combination of pytest-cov, coverage 5 with SQLite and
macOS 10.14 on Azure Pipelines, the pytest process gets killed if we
call os.close(0), which fails the CI. So let's avoid doing that. :)