Conditionally import SO_REUSEADDR, instead of conditionally deleting it
We only expose trio.socket.SO_REUSEADDR on non-Windows platforms.

We used to import it everywhere, and then delete it again on Windows.
But apparently the latest versions of Jedi have grown smart enough to
detect this deletion (but not its conditional nature), so they stopped
being able to complete SO_REUSEADDR.

Out-smart Jedi by skipping importing it on Windows, instead.