add .envrc and shell.nix for direnv+Nix-based development
With this .envrc and shell.nix, merely cd-ing into the trio directory
sets up PATH and PYTHONPATH local to that directory, and runs setup.py
develop, such that one can do trio development with all dependencies
provided from Nix.

The shell.nix is concise because it just takes advantage of the
existing package for trio in Nixpkgs.

I don't expect this commit to be accepted; just submitting this for
reference.