DOC: Demphasize `asks` from awesome libraries (broken)
The project seem to have no activity since October, and does not work
out of the box when installing:

 >>> ImportError: cannot import name 'aopen' from 'anyio'

It requires anyio ~2.x but is actually failing with it.

As it is the first of the list; it is likely to be tried first and users
wondering if they are doing something wrong (guilty of this myself), or
if trio is broken.