Fix documentation references to `socket.socket`
This is a class in the stdlib, not a function. I didn't really dig
into why/when sphinx decided to get fussy about it.

I tested this by building the docs and clicking the link on one of the
changed sites, and it took me to the right place in the Python stdlib docs