subprocess - Fix a bug where trio isn't notified about pidfd closes
If you have
  1. Task A blocks in `Process.wait`
  2. The child process exits
  3. Task B calls `Process.returncode`

(3) will close the pidfd, but nothing will ever wake up Task A, since
trio wasn't notified.

I noticed this when I wrote a `deliver_cancel` function that was slow
enough that the child actually exited before the call to
`Process.returncode`. In the default `_posix_deliver_cancel`, everything is so
fast that the usual ordering is that (3) comes before (2), and there
isn't an early call to `_pidfd_close`.

This fixes the bug by ensuring we notify trio about closing the pidfd
when we do it!