Guard rail cancellable argument of thread dispatch
The cancellable argument is typed bool but at runtime will work with any object with truthiness. However, if someone passes something like an numpy ndarray, with no notion of truthiness, the error is raised in an abort function and causes a TrioInternalError.

There are other unguarded uses of user arguments in abort functions, but they are in the lowlevel namespace, and aren't expected to have robust guardrails.