Fix finding pthread_*name_np on vanilla musl libc
Fix the `pthread_getname_np` and `pthread_setname_np` search logic
to support vanilla versions of musl and CPython, e.g. as used on Gentoo
musl systems.  On such systems, there is no "libpthread.so" (there is
only a static library) and the relevant functions are found
in "libc.so".  Additionally, `ctypes.util.find_library("c")` does not
work because of an old unsolved bug in CPython (linked in the code).

To resolve the problem, add a fallback to trying `libc.so` if no pthread
library can be found.  This roughly covers three possibilities:

- a "typical" system with `libpthread.so` will find that library
  and use it

- a musl system will fall back to `libc.so`, load that library and find
  pthread functions there

- any other system will try to load `libc.so`, and fail

The code in `get_os_thread_name_func()` remains fully relaxed, allowing
either CDLL construction (i.e. finding the library) to fail,
or the library not to contain `pthread_setname_np`.

The code in `test_threads.py` was made more relaxed — rather than
skipping if `libpthread.so` does not exist, it tries to load `libc.so`
as a fallback, and skips if that fails.

Originally reported as https://bugs.gentoo.org/923257.