Remove incorrect `str` possibility from HostnameResolver.getaddrinfo
`_socket.getaddrinfo` takes care of encoding all passed in `str` host
arguments before passing it to a custom resolver. Additionally, the
docs say that a resolver will always receive IDNA-encoded bytes.

However, the type hints of `HostnameResolver.getaddrinfo` imply that a
resolver may need to contend with a `str`

Remove the extraneous appearance