Add type annotations, fixing some issues along the way.
Using eval only worked by accident, because "locals" was
being passed into the "globals" value. InteractiveInterpreter
in the stdlib correctly uses exec, but this doesn't work for
us, because we need to get the values of the expression and
check if it is a coroutine that need to be awaited.

asyncio.__main__ uses the same type.FunctionType idea, which
I originally avoided because I didn't really understand it
and thought eval was simpler...