Drop 3.9 support (#3345)
* Drop 3.9

* Fixes for CI

* Address PR feedback

* [pre-commit.ci] auto fixes from pre-commit.com hooks

for more information, see https://pre-commit.ci

* Better pass through

 - Fix a few typos
 - Remove unnecessary generic function infrastructure
 - Deduplicate `from typing import ...` imports
 - Fix codecov complains

* Undo ruff autofix

* A few more codecov changes

I'm not so sure about the unix pipes change. I suppose we'll get
comments if it doesn't work? I can't tell why the code was there
to begin with, which is a bad sign for being able to tell whether
it's necessary anymore.

* Fix pip-compile files

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>